% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{matpow}
\alias{matpow}
\title{Generalized matrix power function
Implemented by Daniel Kapla
(borrowed from my tvRRR package)}
\usage{
matpow(A, pow, symmetric = FALSE, tol = 1e-07)
}
\description{
Generalized matrix power function
Implemented by Daniel Kapla
(borrowed from my tvRRR package)
}
