% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.rankLME}
\alias{plot.rankLME}
\title{Diagostic plots for rankLME object}
\usage{
\method{plot}{rankLME}(x, which = 2)
}
\arguments{
\item{x}{the rankLME object}

\item{which}{type of plot, see Details}
}
\description{
Provides different diagnostic plots to identify unusual observations
}
\details{
The function creates different diagnostic plots for the rankLME model. The type of
plot is specified by the \code{which} argument:
\itemize{
\item \code{1} Plot of index vs. y by group color
\item \code{2} Plot of fitted values vs. conditional residuals
\item \code{3} Plot of overall leverage vs. standardized conditional residuals
\item \code{4} Plot of groupwise leverage vs. standardized conditional residuals, faceted by group
\item \code{5} Boxplots of the marginal residuals by group
\item \code{6} Variance diagnostic
}
}
