% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pv2}
\alias{pv2}
\title{PV Dataset 2}
\format{
\subsection{\code{pv2}}{

Dataset has the format 42 x 12 and contains data from accelerated aging experiments on photovoltaic modules.
The modules were exposed to two different climatic stresses, denoted by Tropical 1 and Tropical 2.  The data set contains the
power at maximum power point (pmpp) for 6 modules as well as three peak areas extracted from an FTIR spectrum.
}
}
\source{
Analyzed as part of the Austrian Energy Research Program project ADVANCE!
(No. 881133), funded by the Austrian Climate and Energy Fund and the Austrian Research Promotion Agency (FFG).
}
\usage{
pv2
}
\description{
PV Dataset 2
}
\references{
Berger, K., Knöbl, K., Brune, B., Ortner, I., Grubner, S., Oreski, G., Barretta, C.,
Kubicek, B., Neumaier, L., Feichtner, M., Voronko, Y., and Eder, G. (2021). Advanced Degradation Modelling of
Photovoltaic Modules and Materials. 38th European Photovoltaic Solar Energy Conference and Exhibition, pages 745–748, DOI:
\url{10.4229/EUPVSEC20212021-4AV.1.19.}
}
\keyword{datasets}
