% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranklm.R
\name{ranklm}
\alias{ranklm}
\title{RANK-BASED REGRESSION}
\usage{
ranklm(
  X,
  y,
  intercept = TRUE,
  weighted = FALSE,
  leverage_columns = 1:ncol(X),
  V = NULL,
  weights = NULL,
  c = qchisq(0.95, df = ncol(X)),
  mean_function = "hodges_lehmann",
  mean_function_arguments = list(),
  weighting_type = "malahanobis",
  mcd = TRUE,
  ...
)
}
\arguments{
\item{X}{regression matrix}

\item{y}{target}

\item{intercept}{logical, indicates whether an intercept should be modeled}

\item{weighted}{logical, should leverage weights be applied}

\item{leverage_columns}{indicates the columns of X that should be used for calculation
of the leverage weights (ignored if weighted = FALSE)}

\item{weighting_type}{Should malahanobis type weights be used or hat matrix based weights (\code{malahanobis} or \code{hat_matrix})}

\item{mcd}{should MCD be used for covariance estimation or classic estimates (might be necessary for very small groups)}

\item{...}{additional arguments that may be passed to the norm}
}
\value{
The vector of coefficients
}
\description{
Fitting the rank based linear model with starting values given by lm
Input:
}
