% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranklme.R
\name{ranklme}
\alias{ranklme}
\title{Fit the rank-based model}
\usage{
ranklme(
  X,
  y,
  Z,
  g,
  maxit = 10,
  tol = 1e-06,
  intercepts = list(fixed = TRUE, random = TRUE),
  adjust_re = TRUE,
  mean_function = "hodges_lehmann",
  mean_function_arguments_fixed = list(),
  mean_function_arguments_random = list(),
  sd_function = "Qn_corrected",
  sd_function_arguments_fixed = list(),
  sd_function_arguments_random = list(),
  weighted = FALSE,
  use_outlyingness_weights = TRUE,
  ...
)
}
\arguments{
\item{X}{(n x p) matrix of regressors, X does not contain an intercept}

\item{y}{(n x 1) target}

\item{Z}{(n x k) matrix with variables corresponding to random effects, usually a subset
of X; also no intercept}

\item{g}{(n x 1) vector of group matchings for the observations}

\item{maxit}{maximum number of iterations}

\item{tol}{tolerance until convergence}

\item{intercepts}{a named list of length two with elements \code{fixed}  and \code{random}
indicate whether intercepts should be fitted or not}

\item{adjust_re}{logical, indicates whether the random effects should be adjusted for the mean}

\item{mean_function}{string, mean function that should be used for centering}

\item{sd_function}{string, scale estimator that should be used for scaling}
}
\value{
A named list with lots of elements.
}
\description{
Input:
}
