% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{residuals.rankLME}
\alias{residuals.rankLME}
\title{\code{residuals} method for rankLME}
\usage{
\method{residuals}{rankLME}(object, type = "conditional", ...)
}
\description{
\code{residuals} method for rankLME
}
