% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{wilcox_norm_weighted}
\alias{wilcox_norm_weighted}
\title{Weighted Wilcoxon norm, following 2004 Paper by McKean}
\usage{
wilcox_norm_weighted(
  X,
  y,
  beta,
  weighted = TRUE,
  c = qchisq(0.95, df = ncol(X)),
  weights = NULL,
  V = NULL
)
}
\description{
Weighted Wilcoxon norm, following 2004 Paper by McKean
}
\examples{
X <- replicate(4, rnorm(100))
beta <- c(1, 2, 3, 4)
y <- X \%*\% beta + rnorm(100)
X[1:10, ] <- 50
wilcox_norm_weighted(X, y, beta, weighted = TRUE)
}
