% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranklm.R
\name{ranklm}
\alias{ranklm}
\title{RANK-BASED REGRESSION}
\usage{
ranklm(
  X,
  y,
  intercept = TRUE,
  weighted = FALSE,
  V = NULL,
  c = qchisq(0.95, df = ncol(X)),
  mean_function = "hodges_lehmann",
  mean_function_arguments = list(),
  ...
)
}
\arguments{
\item{X}{regression matrix}

\item{y}{target}

\item{intercept}{logical, indicates whether an intercept should be modeled}

\item{...}{additional arguments that may be passed to the norm}

\item{boot}{logical, indicates whether the bootstrap distribution should be
calculated}

\item{n.boot}{number of bootstrap replications}
}
\value{
The vector of coefficients
}
\description{
Fitting the rank based linear model with starting values given by lm
Input:
}
