% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{wilcox_norm_weighted}
\alias{wilcox_norm_weighted}
\title{Weighted Wilcoxon norm, following 2004 Paper by McKean}
\usage{
wilcox_norm_weighted(
  X,
  y,
  beta,
  weighted = TRUE,
  c = qchisq(0.95, df = ncol(X)),
  weights = NULL,
  V = NULL
)
}
\description{
Weighted Wilcoxon norm, following 2004 Paper by McKean
}
